+++ game/shapes.cpp
@@ -4,7 +4,7 @@
 #include <glad/glad.h>
 #include <vector>
 
-void DrawShape(const glm::vec2* data, uint32_t size, uint32_t mode, const ColorType& color, float thickness) {
+void DrawShape(const glm::vec2* data, size_t size, uint32_t mode, const ColorType& color, float thickness) {
 	Resources().GetShader(Shaders::Shapes).Select();
 	Resources().GetShader(Shaders::Shapes).SetVec4("Color", RGBA(color));
 
@@ -65,7 +65,7 @@ void CircleShape::Draw() {
 	LineStripShape poly;
 	poly.Color = Color;
 	poly.Thickness = Thickness;
-	size_t count = Radius;
+	size_t count = static_cast<size_t>(Radius);
 	poly.Points.resize(count);
 	double cnst = 2 * PI / (count - 1);
 	for (size_t i = 0; i < count; i++) {
